GPIO: fix to avoid crash when ctrl register is accessed
In GPIO module, in a bank if none of the pins are in use the
interface clock to that bank will be cut. In the test code make
sure that GPIO ctrl register is not accessed in such scenario.

The patch also corrects the bank_address array & ctrl register
offset address

Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>