Audio-ALSA: Add OMAP5 test cases
Add initial support for OMAP5 audio test cases. It's mostly based on OMAP
android 3.4 kernel:
http://git.omapzoom.org/?p=kernel/omap.git;a=shortlog;h=refs/heads/p-android-omap-3.4

Most of the test cases should work in previous kernels (3.0 or 3.1) for either
OMAP4 or OMAP5.

Some test cases require omapconf and needs to be pushed into the target filesystem
separately in a directory exported in $PATH.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>