AudioALSA: Support ABE DL1-only mode
When ABE is configured in DL1-only mode, the DL2 gain controls and
equalizers aren't available, so the DL1 gains and DL1 Equalizer has
to be shared when handsfree output is enabled.

Due to above reasons, the frequency of sine wave use for testing is
changed to prevent DL1 Equalizer filters that frequency out. On the
volume controls front, the DL1 gains are used instead.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>