ARM: OMAP5: DDR3: Change io settings
The DDR DQ lines are enabled with weak pull. So the DQ line was not staying at Vref
when IDLE (retreats to ground) and because of this there were extra transitions
and noise. So change from 0x64656465 to 0x64646464 to remove the weak pull.

Also internal VREF_DQOUT is set to 0. This has to enabled as well.

With the above two changes better memory stability was observed with extended
temperature ranges around 100C

Signed-off-by: Sricharan R <r.sricharan@ti.com>