am335x_evm: Add check for non-0 profile on GP EVM, update fdtfile
Until we have device tree overlay support or something similar we need to
pass profile-specific device trees to the kernel.  Update the logic to
expose a profile_number variable and update findfdt to use
am335x-evm-profileN.dtb.  This is a TI SDK specific change as the answer
for upstream is to work towards device tree overlays or a similar
concept.

Signed-off-by: Tom Rini <trini@ti.com>