ARM: AM43xx: Do not enable RTC for AM4372 SoC
Commit "am335x: Enable RTC 32K OSC clock" describes the dependency
to enable RTC clks in bootloader. This is not true for AM4372.
In EPOS EVM RTC is not powered (VDDS_RTC grounded to 0). In GP EVM no
need to enble RTC in bootloader. So moving RTC enbling to its respective clock file.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>