ARM: AM4372: Enable caches at starting of u-boot
This is a known issue on AM4372 that when there is a burst read to a
non-cacheable EMIF address space and the burst crosses 1K address boundary will
result in a hang. Since U-boot runs from DDR, there is a possibility that above
case occurs. So enable caches at the beginning of U-boot.
*This is a temporary fix and not meant for mainline.*

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>