am33xx: Add support for modifying INT_CONFIG/PBBPR in EMIF
In EMIF4 blocks of AM335x/TI81XX there is a register at 0x54 called
INT_CONFIG/PBBPR that has a field called PR_OLD_CONFIG that can be
changed depending on workloads of the system to ensure that accesses to
some areas don't cause accesses to other areas to be "stalled".  This
can be seen for example as screen jitter when playing videos.

Signed-off-by: Tom Rini <trini@ti.com>