am335x/omap3_beagle/omap4/omap5: Re-enable fdt_high relocation
The fdt_high variable controls how high into memory the FDT can be moved
as part of booting the kernel.  We had been disabling this feature as by
default we move to the very top of memory which can often be part of
highmem and so not visible to the kernel yet.  However, in other cases
the kernel BSS can overwrite the FDT at the location we use, and we
wouldn't detect this case.  The answer is to re-enable relocation, but
ensure it will be in kernel-visible memory still.

Signed-off-by: Tom Rini <trini@ti.com>