arm: ti-armv7-common: Fix NAND and MMC boot command
When porting the NAND and MMC boot CMD it was developer
error not porting over the contents of the NAND and MMC
boot commands as they existed in the tree.

Therefore need to update the common boot commands to what
was already available for the platforms.

Also removed the NANDARGS from the platform files so that
they do not cause confusion.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
[trini: Add omap5_common.h change]
Signed-off-by: Tom Rini <trini@ti.com>